--- a/log/log.go
+++ b/log/log.go
@@ -65,7 +65,7 @@ type Context struct {
 	logger     Logger
 	keyvals    []interface{}
 	hasValuer  bool
-	projection *deferrableProjection
+	projection Projection
 }
 
 // Log replaces all value elements (odd indexes) containing a Valuer in the
@@ -75,20 +75,15 @@ func (l *Context) Log(keyvals ...interface{}) error {
 	if l == nil {
 		return nil
 	}
-	applyDeferredProjection := false
 	if l.projection != nil {
-		if l.projection.deferred {
-			applyDeferredProjection = true
-		} else {
-			if len(keyvals)%2 != 0 {
-				keyvals = append(keyvals, ErrMissingValue)
-			}
-			projected, preserve := l.projection.Projection(keyvals)
-			if !preserve {
-				return nil
-			}
-			keyvals = projected
+		if len(keyvals)%2 != 0 {
+			keyvals = append(keyvals, ErrMissingValue)
 		}
+		projected, preserve := l.projection(keyvals)
+		if !preserve {
+			return nil
+		}
+		keyvals = projected
 	}
 	kvs := append(l.keyvals, keyvals...)
 	if len(kvs)%2 != 0 {
@@ -102,17 +97,6 @@ func (l *Context) Log(keyvals ...interface{}) error {
 		}
 		bindValues(kvs[:len(l.keyvals)])
 	}
-	if applyDeferredProjection {
-		projected, preserve := l.projection.Projection(kvs)
-		if !preserve {
-			return nil
-		}
-		if len(projected)%2 != 0 {
-			kvs = append(projected, ErrMissingValue)
-		} else {
-			kvs = projected
-		}
-	}
 	return l.logger.Log(kvs...)
 }
 
@@ -121,11 +105,11 @@ func (l *Context) With(keyvals ...interface{}) *Context {
 	if len(keyvals) == 0 || l == nil {
 		return l
 	}
-	if l.projection != nil && !l.projection.deferred {
+	if l.projection != nil {
 		if len(keyvals)%2 != 0 {
 			keyvals = append(keyvals, ErrMissingValue)
 		}
-		projected, preserve := l.projection.Projection(keyvals)
+		projected, preserve := l.projection(keyvals)
 		if !preserve {
 			return nil
 		}
@@ -137,10 +121,11 @@ func (l *Context) With(keyvals ...interface{}) *Context {
 	}
 	return &Context{
 		logger: l.logger,
-		// Limiting the capacity of the stored keyvals ensures that a new
-		// backing array is created if the slice must grow in Log or With.
-		// Using the extra capacity without copying risks a data race that
-		// would violate the Logger interface contract.
+		// Limiting the capacity of the stored keyvals ensures that a
+		// new backing array is created if the slice must grow in Log,
+		// With, or WithProjection. Using the extra capacity without
+		// copying risks a data race that would violate the Logger
+		// interface contract.
 		keyvals:    kvs[:len(kvs):len(kvs)],
 		hasValuer:  l.hasValuer || containsValuer(keyvals),
 		projection: l.projection,
@@ -153,30 +138,34 @@ func (l *Context) WithPrefix(keyvals ...interface{}) *Context {
 	if len(keyvals) == 0 || l == nil {
 		return l
 	}
-	if l.projection != nil && !l.projection.deferred {
+	if l.projection != nil {
 		if len(keyvals)%2 != 0 {
 			keyvals = append(keyvals, ErrMissingValue)
 		}
-		projected, preserve := l.projection.Projection(keyvals)
+		projected, preserve := l.projection(keyvals)
 		if !preserve {
 			return nil
 		}
 		keyvals = projected
 	}
 	// Limiting the capacity of the stored keyvals ensures that a new
-	// backing array is created if the slice must grow in Log or With.
-	// Using the extra capacity without copying risks a data race that
-	// would violate the Logger interface contract.
+	// backing array is created if the slice must grow in Log, With,
+	// or WithProjection. Using the extra capacity without copying
+	// risks a data race that would violate the Logger interface
+	// contract.
 	n := len(l.keyvals) + len(keyvals)
+	valueMissing := false
 	if len(keyvals)%2 != 0 {
+		valueMissing = true
 		n++
 	}
-	kvs := make([]interface{}, 0, n)
-	kvs = append(kvs, keyvals...)
-	if len(kvs)%2 != 0 {
-		kvs = append(kvs, ErrMissingValue)
+	kvs := make([]interface{}, n)
+	base := copy(kvs, keyvals)
+	if valueMissing {
+		kvs[base] = ErrMissingValue
+		base++
 	}
-	kvs = append(kvs, l.keyvals...)
+	copy(kvs[base:], l.keyvals)
 	return &Context{
 		logger:     l.logger,
 		keyvals:    kvs,
@@ -194,80 +183,40 @@ func (l *Context) WithPrefix(keyvals ...interface{}) *Context {
 // number of keys and values paired in an alternating sequence.
 type Projection func(keyvals []interface{}) (kvs []interface{}, preserve bool)
 
-type deferrableProjection struct {
-	Projection
-	deferred bool
-}
-
-// WithPreProjection returns a new Context that applies the supplied
+// WithProjection returns a new Context that applies the supplied
 // projection function before any other such functions already bound
 // to the original Context.
-//
-// If deferred, it waits to apply the function only to complete log
-// records before submitting them first to any other projection
-// functions and then to its downstream logger. Otherwise, it applies
-// the function to keys and values as they accumulate in calls to
-// With, WithPrefix, and Log, before interpolating any contributions
-// from Valuers.
-func (l *Context) WithPreProjection(p Projection, deferred bool) *Context {
+func (l *Context) WithProjection(p Projection) *Context {
 	if p == nil || l == nil {
 		return l
 	}
 	if preceding := l.projection; preceding != nil {
-		if !deferred && preceding.deferred {
-			deferred = true
-		}
 		inner := p
-		outer := preceding.Projection
 		p = func(keyvals []interface{}) ([]interface{}, bool) {
 			kvs, preserve := inner(keyvals)
 			if !preserve {
 				return nil, false
 			}
-			return outer(kvs)
+			return preceding(kvs)
 		}
+	}
+	projected, preserve := p(l.keyvals)
+	if !preserve {
+		return nil
+	}
+	if len(projected)%2 != 0 {
+		projected = append(projected, ErrMissingValue)
 	}
 	return &Context{
-		logger:     l.logger,
-		keyvals:    l.keyvals,
+		logger: l.logger,
+		// Limiting the capacity of the stored keyvals ensures that a
+		// new backing array is created if the slice must grow in Log,
+		// With, or WithProjection. Using the extra capacity without
+		// copying risks a data race that would violate the Logger
+		// interface contract.
+		keyvals:    projected[:len(projected):len(projected)],
 		hasValuer:  l.hasValuer,
-		projection: &deferrableProjection{p, deferred},
-	}
-}
-
-// WithPostProjection returns a new Context that applies the supplied
-// projection function after any other such functions already bound
-// to the original Context.
-//
-// If deferred, it waits to apply the function only to complete log
-// records as returned from any other projection functions, and
-// submits the result to its downstream logger. Otherwise, it applies
-// the function to keys and values as they accumulate in calls to
-// With, WithPrefix, and Log, before interpolating any contributions
-// from Valuers.
-func (l *Context) WithPostProjection(p Projection, deferred bool) *Context {
-	if p == nil || l == nil {
-		return l
-	}
-	if preceding := l.projection; preceding != nil {
-		if !deferred && preceding.deferred {
-			deferred = true
-		}
-		outer := p
-		inner := preceding.Projection
-		p = func(keyvals []interface{}) ([]interface{}, bool) {
-			kvs, preserve := inner(keyvals)
-			if !preserve {
-				return nil, false
-			}
-			return outer(kvs)
-		}
-	}
-	return &Context{
-		logger:     l.logger,
-		keyvals:    l.keyvals,
-		hasValuer:  l.hasValuer,
-		projection: &deferrableProjection{p, deferred},
+		projection: p,
 	}
 }