Following up from #444, demonstrate an alternate approach to the "level" package offered via the "github.com/go-kit/kit/log/experimental_package" import path (pending promotion via #449), allowing log event level filtering via functions that specify the filtering threshold.

This capability is implemented in terms of a new concept for `log.Context`: _projection functions_. There are two new exported methods on `log.Context` that mention a new exported type, though defining that type is a convenience and isn't strictly necessary:
- **`log.Projection`**  
  A convenience declaration of a _projection function_ that accepts a proposed set of key/value pairs and either drops the record (and any later records that would include those key/value pairs) or proposes a revised set of key/value pairs.
- **`log.Context.WithPreProjection`**  
  Composes a _projection function_ onto a `Context`, applying it early as the "innermost" such function.
- **`log.Context.WithPostProjection`**  
  Composes a _projection function_ onto a `Context`, applying it late as the "outermost" such function.

Including this function in `log.Context` increases the struct's size by four bytes, which accounts for the 16.67% increase in allocations per operation seen in the benchmarks. That is, each time we allocate a new `Context` in `log.NewContext`, we allocate 28 instead of 24 bytes.

There are a few operations that are slower than the alternate operation, such as the JSON and _logfmt_ conversion. I _think_ this is due to my level values being implementations of `fmt.Stringer` as opposed to `string` values; both those formatters visit such values second in line, and call on them with a deferred function to trap panics. As my interest here is more in cutting down on CPU and allocation cost in cases where most log messages are filtered out, I stuck with my `level.levelValue` type here, as it yielded faster benchmarks for discerning whether the value should be dropped by the filter.

With these _projection functions_ in place, it's possible to implement [log15's `ext.EscalateErrHandler`](https://godoc.org/github.com/inconshreveable/log15/ext#EscalateErrHandler), though only within the `level` package, because the `level.levelValue` type is not exported. We would have to make a copy of the supplied key/value slice and replace the level value in it if we discover a non-nil `error` elsewhere in the sequence, so it's not as efficient as `ext.EscalateErrHandler`, but that library's records include the level directly.

Invoking these projection functions involves trading trust for efficiency. I considered making a defensive copy of the key/value sequence supplied to each invocation of the function, which allows the function the opportunity to mutate the sequence in place. However, that induces allocation and copying even for cases where the projection function is just a pure filter; even sequences it would vote to drop would have been copied first. Instead, I decided to document that a well-behaved `log.Projection` does not mutate the supplied slice, expecting that the most common cases would involve only inspecting the proposed key/value pairs and either dropping them or preserving them as they are. I expect that this choice will be controversial.

Here are the benchmark comparisons:

| name  | old time/op | new time/op  | delta |
| --- | --- | --- | --- |
| NopBaseline-8           |  470ns ± 0% |  443ns ± 0% |  -5.74% |
| NopDisallowedLevel-8    |  513ns ± 0% |  187ns ± 0% | -63.55% |
| NopAllowedLevel-8       |  516ns ± 0% |  451ns ± 0% | -12.60% |
| JSONBaseline-8          | 2.44µs ± 0% | 2.40µs ± 0% |  -1.88% |
| JSONDisallowedLevel-8   |  515ns ± 0% |  184ns ± 0% | -64.27% |
| JSONAllowedLevel-8      | 2.48µs ± 0% | 2.34µs ± 0% |  -5.77% |
| LogfmtBaseline-8        | 1.05µs ± 0% | 1.13µs ± 0% |  +7.71% |
| LogfmtDisallowedLevel-8 |  514ns ± 0% |  181ns ± 0% | -64.79% |
| LogfmtAllowedLevel-8    | 1.13µs ± 0% | 1.13µs ± 0% |  -0.18% |

| name | old alloc/op | new alloc/op | delta |
| --- | --- | --- | --- |
| NopBaseline-8           | 288B ± 0% |  336B ± 0% | +16.67% |
| NopDisallowedLevel-8    | 288B ± 0% |  112B ± 0% | -61.11% |
| NopAllowedLevel-8       | 288B ± 0% |  336B ± 0% | +16.67% |
| JSONBaseline-8          | 968B ± 0% | 1032B ± 0% |  +6.61% |
| JSONDisallowedLevel-8   | 288B ± 0% |  112B ± 0% | -61.11% |
| JSONAllowedLevel-8      | 968B ± 0% | 1032B ± 0% |  +6.61% |
| LogfmtBaseline-8        | 288B ± 0% |  336B ± 0% | +16.67% |
| LogfmtDisallowedLevel-8 | 288B ± 0% |  112B ± 0% | -61.11% |
| LogfmtAllowedLevel-8    | 288B ± 0% |  336B ± 0% | +16.67% |

| name | old allocs/op | new allocs/op | delta |
| --- | --- | --- | --- |
| NopBaseline-8           | 9.00 ± 0% | 9.00 ± 0% |  +0.00% |
| NopDisallowedLevel-8    | 9.00 ± 0% | 5.00 ± 0% | -44.44% |
| NopAllowedLevel-8       | 9.00 ± 0% | 9.00 ± 0% |  +0.00% |
| JSONBaseline-8          | 22.0 ± 0% | 23.0 ± 0% |  +4.55% |
| JSONDisallowedLevel-8   | 9.00 ± 0% | 5.00 ± 0% | -44.44% |
| JSONAllowedLevel-8      | 22.0 ± 0% | 23.0 ± 0% |  +4.55% |
| LogfmtBaseline-8        | 9.00 ± 0% | 9.00 ± 0% |  +0.00% |
| LogfmtDisallowedLevel-8 | 9.00 ± 0% | 5.00 ± 0% | -44.44% |
| LogfmtAllowedLevel-8    | 9.00 ± 0% | 9.00 ± 0% |  +0.00% |

They are not as obviously rosy. I think that this approach's benefits would become more clear when we see filtering applied to a `log.Context` with lots of log messages dropped, such as the following:
```go
func BenchmarkManyDroppedRecords(b *testing.B) {
	logger := level.AllowingInfoAndAbove(log.NewJSONLogger(ioutil.Discard))
	b.ResetTimer()
	b.ReportAllocs()
	debug := level.Debug(logger)
	info := level.Info(logger)
	for i := 0; i < b.N; i++ {
		debug.Log("foo", "1")
		// Only this one will be retained.
		info.Log("baz", "quux")
		debug.Log("foo", "2")
		debug.Log("foo", "3")
		debug.Log("foo", "4")
	}
}
```
Run against the equivalent version from the _experimental_level_ directory:

| name  | old time/op | new time/op  | delta |
| --- | --- | --- | --- |
| ManyDroppedRecords-8 | 3.36µs ± 0% | 2.79µs ± 0% | -16.80% |

| name | old alloc/op | new alloc/op | delta |
| --- | --- | --- | --- |
| ManyDroppedRecords-8 | 1.32kB ± 0% | 1.08kB ± 0% | -18.18% |

| name | old allocs/op | new allocs/op | delta |
| --- | --- | --- | --- |
| ManyDroppedRecords-8 | 33.0 ± 0% | 30.0 ± 0% | -9.09% |

Lacking still are unit tests to cover the new `Context.WithPreProjection` and `Context.WithPostProjection` functions. As each has two parameters, the first of which can be nil, there are at least four interesting cases for each one, as well as exercising proper composition with an existing projection function in the `Context`. I didn't write those yet, as I am uncertain whether the maintainers are amenable to the ideas introduced here.

I welcome your thoughts, but if you're tired of considering these alternative proposals, I understand.