@seh 2017-02-08
Here's a hastily written implementation of error escalation, situated within the "level" package:
```go
func EscalatingErrors(logger log.Logger) log.Logger {
	return log.NewContext(logger).WithPreProjection(func(keyvals []interface{}) ([]interface{}, bool) {
		resolve := func(v *levelValue, index int) ([]interface{}, bool) {
			if v != valError {
				kvs := make([]interface{}, len(keyvals))
				base := copy(kvs, keyvals[:index])
				kvs[index] = valError
				copy(kvs[base+1:], keyvals[index+1:])
				return kvs, true
			}
			return keyvals, true
		}
		for i, end := 1, len(keyvals); i < end; i += 2 {
			switch v := keyvals[i].(type) {
			case *levelValue:
				index := i
				i += 2
				for ; i < end; i += 2 {
					if err, ok := keyvals[i].(error); ok && err != nil {
						return resolve(v, index)
					}
				}
				break
			case error:
				if v == nil {
					continue
				}
				i += 2
				for ; i < end; i += 2 {
					if l, ok := keyvals[i].(*levelValue); ok {
						return resolve(l, i)
					}
				}
				break
			}
		}
		return keyvals, true
	}, true)
}
```
Note that it's a "pre" (early) projection, to intervene before the "info"-and-above filter, and it's deferred, to preclude the "info"-and-above filter from rendering the `Context` inert as soon as it sees the "debug"-level value captured early on, before we've seen whether there's a non-nil `error` value in the eventual log record.

A corresponding test case:
```go
func TestEscalatingErrors(t *testing.T) {
	var buf bytes.Buffer
	var logger log.Logger
	logger = level.EscalatingErrors(level.AllowingInfoAndAbove(log.NewLogfmtLogger(&buf)))
	logger = level.Debug(logger)
	logger.Log("msg", "success", "err", nil)
	logger.Log("msg", "failure", "err", errors.New("failed"))
	if want, have := `level=error msg=failure err=failed`, strings.TrimSpace(buf.String()); want != have {
		t.Errorf("want %q, have %q", want, have)
	}
}
```