@ChrisHines 2017-02-10
I found some time to read through this idea today. I haven't completely wrapped my head around the new concepts. I get the idea of a projection, but the implementation and its ramifications are not obvious to me after 1.5 passes. That is not encouraging.

From a user's perspective I am concerned that the interaction between pre/post projections are not easy grasp.

From an API design perspective Projections and Loggers are not orthogonal. I think deferred projections are redundant because the same choices could be made in the Logger called at the end of `Context.Log` without any extra overhead. That creates two ways to do common tasks with this package. That is undesirable.