@seh 2017-02-10
Thank you for considering the proposal. I value your feedback.

I started out with just the "pre" projection, and no deferral. I only added the deferral option after you mentioned `ext.EscalateErrorHandler`, and it seemed like it was a variation on what I was in the middle of implementing. It wasn't until early yesterday morning (when all things from the day prior become more clear) that I thought, "You know, that deferred projection is really the same as a `log.Logger`." I didn't go back and remove the deferral option, though, not yet being sure whether it was a worthwhile idea.

As for the pre- and post-composition, I thought of this analogously to languages like Common Lisp that have the notion of "before methods," "after methods," and "around methods." I didn't want repeated calls to what was originally `Context.WithProjection` to snuff out preceding projections already bound to the `Context`, but I wasn't sure that it's always correct to treat the new projection function as either the "outer" one or the "inner" one, if we are going to compose a new function with it. For leveled record filtering, calling the filtering function earliest seems correct.

Do you think it's worth pulling back from the deferral and composition options, and just exposing a single `Context.WithProjection` method? I expect that would take me a few hours to whittle down.