@ChrisHines 2017-02-10
Yes, I would be interested to see what the simpler version looks like.