@seh 2017-02-10
Good. I'm already most of the way through it.
I will probably have to wait until Monday to post an update here. I'll provide fresh benchmark results too.