@seh 2017-02-13
Per your suggestion, I retained only `Context.WithProjection`, which is composed before any previous projection functions.

Here are the latest benchmarks:
```
name                          old time/op    new time/op    delta
NopBaseline-8                    518ns ± 0%     485ns ± 0%   -6.37%
NopDisallowedLevel-8             568ns ± 0%     206ns ± 0%  -63.73%
NopAllowedLevel-8                568ns ± 0%     490ns ± 0%  -13.73%
JSONBaseline-8                  2.52µs ± 0%    2.64µs ± 0%   +4.96%
JSONDisallowedLevel-8            563ns ± 0%     202ns ± 0%  -64.12%
JSONAllowedLevel-8              2.57µs ± 0%    2.72µs ± 0%   +5.88%
LogfmtBaseline-8                1.14µs ± 0%    1.28µs ± 0%  +12.00%
LogfmtDisallowedLevel-8          562ns ± 0%     202ns ± 0%  -64.06%
LogfmtAllowedLevel-8            1.20µs ± 0%    1.26µs ± 0%   +5.51%
DroppedRecords/1-of-2-8         1.79µs ± 0%    1.64µs ± 0%   -8.11%
DroppedRecords/3-of-4-8         1.77µs ± 0%    1.65µs ± 0%   -7.21%
DroppedRecords/9-of-10-8        1.77µs ± 0%    1.65µs ± 0%   -6.93%
DroppedRecords/99-of-100-8      1.78µs ± 0%    1.62µs ± 0%   -9.31%
DroppedRecords/999-of-1000-8    1.77µs ± 0%    1.64µs ± 0%   -7.62%

name                          old alloc/op   new alloc/op   delta
NopBaseline-8                     352B ± 0%      336B ± 0%   -4.55%
NopDisallowedLevel-8              352B ± 0%      112B ± 0%  -68.18%
NopAllowedLevel-8                 352B ± 0%      336B ± 0%   -4.55%
JSONBaseline-8                  1.03kB ± 0%    1.03kB ± 0%   +0.00%
JSONDisallowedLevel-8             352B ± 0%      112B ± 0%  -68.18%
JSONAllowedLevel-8              1.03kB ± 0%    1.03kB ± 0%   +0.00%
LogfmtBaseline-8                  352B ± 0%      336B ± 0%   -4.55%
LogfmtDisallowedLevel-8           352B ± 0%      112B ± 0%  -68.18%
LogfmtAllowedLevel-8              352B ± 0%      336B ± 0%   -4.55%
DroppedRecords/1-of-2-8           335B ± 0%      335B ± 0%   +0.00%
DroppedRecords/3-of-4-8           335B ± 0%      335B ± 0%   +0.00%
DroppedRecords/9-of-10-8          335B ± 0%      335B ± 0%   +0.00%
DroppedRecords/99-of-100-8        335B ± 0%      335B ± 0%   +0.00%
DroppedRecords/999-of-1000-8      335B ± 0%      335B ± 0%   +0.00%

name                          old allocs/op  new allocs/op  delta
NopBaseline-8                     10.0 ± 0%       9.0 ± 0%  -10.00%
NopDisallowedLevel-8              10.0 ± 0%       5.0 ± 0%  -50.00%
NopAllowedLevel-8                 10.0 ± 0%       9.0 ± 0%  -10.00%
JSONBaseline-8                    23.0 ± 0%      23.0 ± 0%   +0.00%
JSONDisallowedLevel-8             10.0 ± 0%       5.0 ± 0%  -50.00%
JSONAllowedLevel-8                23.0 ± 0%      23.0 ± 0%   +0.00%
LogfmtBaseline-8                  10.0 ± 0%       9.0 ± 0%  -10.00%
LogfmtDisallowedLevel-8           10.0 ± 0%       5.0 ± 0%  -50.00%
LogfmtAllowedLevel-8              10.0 ± 0%       9.0 ± 0%  -10.00%
DroppedRecords/1-of-2-8           8.00 ± 0%      7.00 ± 0%  -12.50%
DroppedRecords/3-of-4-8           8.00 ± 0%      7.00 ± 0%  -12.50%
DroppedRecords/9-of-10-8          8.00 ± 0%      7.00 ± 0%  -12.50%
DroppedRecords/99-of-100-8        8.00 ± 0%      7.00 ± 0%  -12.50%
DroppedRecords/999-of-1000-8      8.00 ± 0%      7.00 ± 0%  -12.50%
```
It looks like I'll need to rebase atop #449 and #454.