@ChrisHines 2017-02-20
I think we have different goals for what we are trying to measure. I want to know the amortized cost of a `Log` call. For my goal the criticism is fundamental. I don't think you had that goal in mind, so from your perspective I suppose it is more aesthetic. But I'd like to persuade you that it is important because keeping a consistent unit of operation allows deeper analysis. It gives meaning to the `time/op` data, while your approach only has meaning for the `delta` column.

In my benchmark results we can compare the results between both old and new implementations as well as between different ratios within an implementation. In other words, with a consistent "op" size we can compare the times across both rows and columns, not just columns.
```
name                          old time/op    new time/op    delta
DroppedRecords/1-of-2-8         1.93µs ± 1%    1.83µs ± 1%   -5.17%  (p=0.029 n=4+4)
DroppedRecords/3-of-4-8         1.09µs ± 2%    0.97µs ± 1%  -10.71%  (p=0.008 n=5+5)
DroppedRecords/9-of-10-8         591ns ± 3%     467ns ± 2%  -21.07%  (p=0.008 n=5+5)
DroppedRecords/99-of-100-8       297ns ± 6%     155ns ± 4%  -47.84%  (p=0.008 n=5+5)
DroppedRecords/999-of-1000-8     263ns ± 4%     130ns ± 6%  -50.68%  (p=0.008 n=5+5)
```
When I look at these numbers I clearly see that the new implementation filters records faster than the old, but also that the savings gained by filtering a higher ratio of log records dwarf the gains of faster filtering. I can also see that although the percentage difference between old and new is much higher when most log events are discarded, the absolute speed difference is fairly constant across all ratios. That fact tells me that the gains come from a reduction in O(1) overhead rather than a O(n) scaling factor.